% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{dbWriteTable}
\alias{dbWriteTable}
\title{Insert data from a dataframe into a table using the specified DatabaseConnectionManager object.}
\usage{
dbWriteTable(dbManager, table = "", value = NULL, append = TRUE)
}
\arguments{
\item{dbManager}{DatabaseConnectionManager object to run the query against}

\item{table}{a string containing the table name you want to insert into}

\item{value}{a dataframe containing the data to insert}

\item{append}{a boolean indicating whether to append the data or not (optional; defaults to TRUE)}
}
\value{
Nothing
}
\description{
Insert data from a dataframe into a table using the specified DatabaseConnectionManager object.
}
\examples{
dbWriteTable(dbManager, "users")
runQuery(dbManager, useFile = TRUE, queryLocation = "~/query.sql")
}
\seealso{
\code{\link{runQuery}}
}

