% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{destroyConnection}
\alias{destroyConnection}
\title{Clears up the resources used by a DatabaseConnectionManager}
\usage{
destroyConnection(dbManager)
}
\arguments{
\item{dbManager}{DatabaseConnectionManager object to run the query against}
}
\value{
Nothing. Side-effects only.
}
\description{
Clears up resources by disconnecting all the connections to the driver, 
then unloading the driver
}
\examples{
destroyConnection(dbManager)
}
\seealso{
\code{\link{newConnection}}, \code{\link{runQuery}}
}

