% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{runQuery}
\alias{runQuery}
\title{Run a query against a PostgresConnectionManager object}
\usage{
runQuery(pgManager, query = "", useFile = FALSE, queryLocation = "")
}
\arguments{
\item{pgManager}{PostgresConnectionManager object to run the query against}

\item{query}{a string containing query to run against the connection (optional; needed if useFile = FALSE)}

\item{useFile}{a boolean indicating whether to run a query directly or run the query from file (optional; defaults to FALSE)
If TRUE, queryLocation must be present and pointing to an existing file.
If FALSE, query must be specified.}

\item{queryLocation}{a string pointing at the location of the query file (optional; needed if useFile = TRUE)}
}
\value{
The resulting data frame
}
\description{
Runs a query against the specified PostgresConnectionManager object, either by specifying the SQL directly or 
pointing at a file containing the query.
}
\examples{
runQuery(pgManager, "SELECT * FROM users;")
runQuery(pgManager, useFile = TRUE, queryLocation = "~/query.sql")
}
\seealso{
\code{\link{newConnection}}, \code{\link{destroyConnection}}
}

