% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/manager.R
\name{destroyConnection}
\alias{destroyConnection}
\title{Clears up the resources used by a PostgresConnectionManager}
\usage{
destroyConnection(pgManager)
}
\arguments{
\item{pgManager}{PostgresConnectionManager object to run the query against}
}
\value{
Nothing. Side-effects only.
}
\description{
Clears up resources by disconnecting all the connections to the driver,
then unloading the driver
}
\examples{
destroyConnection(pgManager)
}
\seealso{
\code{\link{newConnection}}, \code{\link{runQuery}}
}

